requirements.txt
openai
requests
moviepy
python-dotenv
